+++ PythonTest1/PythonTest1.pyproj
@@ -4,7 +4,7 @@
     <SchemaVersion>2.0</SchemaVersion>
     <ProjectGuid>8747b4b3-c560-461f-8937-a5603a46864b</ProjectGuid>
     <ProjectHome>.</ProjectHome>
-    <StartupFile>Program.py</StartupFile>
+    <StartupFile>run.py</StartupFile>
     <SearchPath>
     </SearchPath>
     <WorkingDirectory>.</WorkingDirectory>
@@ -32,7 +32,7 @@
     <Compile Include="Glitchcore_Generator.py">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Program.py" />
+    <Compile Include="run.py" />
   </ItemGroup>
   <ItemGroup>
     <InterpreterReference Include="Global|PythonCore|3.9" />
